backend/pyproject.toml
[project]
name = "fairlens-backend"
version = "0.1.0"
description = "FairLens AI Bias Auditing Tool — Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pandas",
    "numpy",
    "scikit-learn",
    "fairlearn",
    "scipy",
    "google-genai",
    "python-multipart",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
